+++ CHANGES.rst
@@ -1,6 +1,15 @@
 Changelog
 =========
 
+0.0.4 (2020-06-04)
+------------------
+
+- added option `--ignore_errors` to keep debayering images despite errors
+- added option `--delete` to delete input file when successfully debayered
+- outputting total processing time now at the end
+- split code into more fine-grained methods to make it easier to use the tool as Python library
+
+
 0.0.3 (2020-05-25)
 ------------------